requirements.txt
streamlit
pandas
numpy
requests
openai
pywin32
newspaper3k
scikit-learn
openpyxl
lxml